refactor(approve): hoist SignaturePad to shared components
Trip app needs the same draw/upload signature pad. Move it from
approve/_components into apps/portal/components so both apps import
one source of truth instead of duplicating.
