+++ apps/portal/components/signature-pad.tsx
